2.0
四頁-->兩頁(僅有輸入&輸出)

+++ CMDXMoldTesting/DataCenter.h
@@ -67,7 +67,7 @@ public:
 		return m_iMaterialSel;
 	}
 
-	void SetMaxScrewStroke(int stroke)
+	void SetMaxScrewStroke(double stroke)
 	{
 		m_dMaxScrewStroke = stroke;
 	}
